patch: better internal structure
