+++ src/main/java/conseq4j/SyncListenableExecutorService.java
@@ -20,41 +20,21 @@
 
 package conseq4j;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
 import java.util.concurrent.BlockingQueue;
-import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 
 /**
  * @author Qingitan Wang
  */
-abstract class SyncListenableExecutorService extends ThreadPoolExecutor implements ListenableExecutorService {
+abstract class SyncListenableExecutorService extends ListenableExecutorServiceTemplate {
 
-    protected final List<ExecutorServiceListener> executorServiceListeners = Collections.synchronizedList(
-            new ArrayList<>());
-
-    protected SyncListenableExecutorService(int corePoolSize, int maximumPoolSize, long keepAliveTime, TimeUnit unit,
+    SyncListenableExecutorService(int corePoolSize, int maximumPoolSize, long keepAliveTime, TimeUnit unit,
             BlockingQueue<Runnable> workQueue) {
         super(corePoolSize, maximumPoolSize, keepAliveTime, unit, workQueue);
     }
 
-    abstract void doBeforeExecute(Thread t, Runnable r);
-
-    abstract void doAfterExecute(Runnable r, Throwable t);
-
     @Override
-    protected void beforeExecute(Thread t, Runnable r) {
-        doBeforeExecute(t, r);
-        super.beforeExecute(t, r);
-        if (executorServiceListeners.isEmpty()) {
-            return;
-        }
-        notifyListenersBeforeExecute(t, r);
-    }
-
-    protected void notifyListenersBeforeExecute(Thread t, Runnable r) {
+    void notifyListenersBeforeExecute(Thread t, Runnable r) {
         synchronized (executorServiceListeners) {
             executorServiceListeners.forEach(executorServiceListener -> executorServiceListener.beforeEachExecute(t,
                     r));
@@ -62,29 +42,10 @@ abstract class SyncListenableExecutorService extends ThreadPoolExecutor implemen
     }
 
     @Override
-    protected void afterExecute(Runnable r, Throwable t) {
-        super.afterExecute(r, t);
-        doAfterExecute(r, t);
-        if (executorServiceListeners.isEmpty()) {
-            return;
-        }
-        notifyListenersAfterExecute(r, t);
-    }
-
-    protected void notifyListenersAfterExecute(Runnable r, Throwable t) {
+    void notifyListenersAfterExecute(Runnable r, Throwable t) {
         synchronized (executorServiceListeners) {
             executorServiceListeners.forEach(executorServiceListener -> executorServiceListener.afterEachExecute(r, t));
         }
-    }
-
-    @Override
-    public void addListener(ExecutorServiceListener executorServiceListener) {
-        executorServiceListeners.add(executorServiceListener);
-    }
-
-    @Override
-    public void clearListeners() {
-        this.executorServiceListeners.clear();
     }
 
 }
